project/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
greenlet==3.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
